+++ src/path/Path.js
@@ -291,7 +291,8 @@ var Path = PathItem.extend(/** @lends Path# */{
         }
     }
 }, /** @lends Path# */{
-    // Enforce bean creation for getPathData(), as it has hidden parameters.
+    // Enforce bean creation for getPathData() and getArea(), as they have
+    // hidden parameters.
     beans: true,
 
     getPathData: function(_matrix, _precision) {
@@ -348,8 +349,7 @@ var Path = PathItem.extend(/** @lends Path# */{
             parts.push('z');
         }
         return parts.join('');
-    }
-}, /** @lends Path# */{
+    },
 
     // TODO: Consider adding getSubPath(a, b), returning a part of the current
     // path, with the added benefit that b can be < a, and closed looping is
@@ -820,19 +820,25 @@ var Path = PathItem.extend(/** @lends Path# */{
      * @bean
      * @type Number
      */
-    getArea: function() {
-        if (this._area == null) {
+    getArea: function(_closed) {
+        // If the call overrides the 'closed' state, do not cache the result.
+        // This is used in tracePaths().
+        var cached = _closed === undefined,
+            area = this._area;
+        if (!cached || area == null) {
             var segments = this._segments,
                 count = segments.length,
-                last = count - 1,
-                area = 0;
-            for (var i = 0, l = this._closed ? count : last; i < l; i++) {
+                closed = cached ? this._closed : _closed,
+                last = count - 1;
+            area = 0;
+            for (var i = 0, l = closed ? count : last; i < l; i++) {
                 area += Curve.getArea(Curve.getValues(
                         segments[i], segments[i < last ? i + 1 : 0]));
             }
-            this._area = area;
+            if (cached)
+                this._area = area;
         }
-        return this._area;
+        return area;
     },
 
     /**
@@ -1092,15 +1098,19 @@ var Path = PathItem.extend(/** @lends Path# */{
         return location != null ? this.splitAt(location) : null;
     },
 
-    // DOCS: document Path#join(path) in more detail.
-    // DOCS: document Path#join() (joining with itself)
-    // TODO: Consider adding a distance / tolerance parameter for merging.
     /**
-     * Joins the path with the specified path, which will be removed in the
-     * process.
+     * Joins the path with the other specified path, which will be removed in
+     * the process. They can be joined if the first or last segments of either
+     * path lie in the same location. Locations are optionally compare with a
+     * provide `tolerance` value.
      *
-     * @param {Path} path the path to join this path with
-     * @return {Path} the joined path
+     * If `null` or `this` is passed as the other path, the path will be joined
+     * with itself if the first and last segment are in the same location.
+     *
+     * @param {Path} path the path to join this path with; `null` or `this` to
+     *     join the path with itself
+     * @param {Number} [tolerance=0] the tolerance with which to decide if two
+     *     segments are to be considered the same location when joining
      *
      * @example {@paperscript}
      * // Joining two paths:
@@ -1161,25 +1171,26 @@ var Path = PathItem.extend(/** @lends Path# */{
      * // Select the path to show that they have joined:
      * path.selected = true;
      */
-    join: function(path) {
-        if (path) {
+    join: function(path, tolerance) {
+        var epsilon = tolerance || 0;
+        if (path && path !== this) {
             var segments = path._segments,
                 last1 = this.getLastSegment(),
                 last2 = path.getLastSegment();
             if (!last2) // an empty path?
                 return this;
-            if (last1 && last1._point.equals(last2._point))
+            if (last1 && last1._point.isClose(last2._point, epsilon))
                 path.reverse();
             var first2 = path.getFirstSegment();
-            if (last1 && last1._point.equals(first2._point)) {
+            if (last1 && last1._point.isClose(first2._point, epsilon)) {
                 last1.setHandleOut(first2._handleOut);
                 this._add(segments.slice(1));
             } else {
                 var first1 = this.getFirstSegment();
-                if (first1 && first1._point.equals(first2._point))
+                if (first1 && first1._point.isClose(first2._point, epsilon))
                     path.reverse();
                 last2 = path.getLastSegment();
-                if (first1 && first1._point.equals(last2._point)) {
+                if (first1 && first1._point.isClose(last2._point, epsilon)) {
                     first1.setHandleIn(last2._handleIn);
                     // Prepend all segments from path except the last one.
                     this._add(segments.slice(0, segments.length - 1), 0);
@@ -1197,7 +1208,7 @@ var Path = PathItem.extend(/** @lends Path# */{
         // only if its ends touch.
         var first = this.getFirstSegment(),
             last = this.getLastSegment();
-        if (first !== last && first._point.equals(last._point)) {
+        if (first !== last && first._point.isClose(last._point, epsilon)) {
             first.setHandleIn(last._handleIn);
             last.remove();
             this.setClosed(true);
@@ -2508,10 +2519,9 @@ new function() { // PostScript-style drawing commands
             }
         },
 
-        closePath: function(join) {
+        closePath: function(tolerance) {
             this.setClosed(true);
-            if (join)
-                this.join();
+            this.join(this, tolerance);
         }
     };
 }, { // A dedicated scope for the tricky bounds calculations